// src/app/models/product.ts
export class Product {
    id: number;
    category: number;
    subcategory: number;
    productName: string;
    productCompany: string;
    productPrice: number;
    productPriceBeforeDiscount: number;
    productDescription: string;
    productImage1: string;
    productImage2: string;
    productImage3: string;
    shippingCharge: number;
    productAvailability: string;
    postingDate: Date;
    updationDate: Date;
    priceVarietyAvailable = false;
}